+++ frontend/lost/src/containers/pipeline/running/modals/types/DatasourceModal.jsx
@@ -1,40 +1,51 @@
-import CollapseCard from '../../../globalComponents/modals/CollapseCard'
-import Table from '../../../globalComponents/modals/Table'
-import { CModalBody, CModalHeader } from '@coreui/react'
+import { createColumnHelper } from '@tanstack/react-table'
+import CoreDataTable from '../../../../../components/CoreDataTable'
+import BaseModal from '../../../../../components/BaseModal'
 
-const DatasourceModal = (props) => {
+const DatasourceModal = ({ datasource, state, id, modalOpened, onClose }) => {
+  const columnHelper = createColumnHelper()
+  const columns = [
+    columnHelper.accessor('taskName', {
+      header: () => 'Attribute',
+      cell: ({ row }) => {
+        return <b>{row.original.key}</b>
+      },
+    }),
+    columnHelper.accessor('taskName', {
+      header: () => 'Value',
+      cell: ({ row }) => {
+        return <>{row.original.value}</>
+      },
+    }),
+  ]
+  const tData = [
+    {
+      key: 'Path:',
+      value: datasource.rawFilePath,
+    },
+    {
+      key: 'Element ID:',
+      value: id,
+    },
+    {
+      key: 'Datasource ID:',
+      value: datasource.id,
+    },
+    {
+      key: 'Status:',
+      value: state,
+    },
+  ]
+  console.log('Should be true', modalOpened)
   return (
-    <>
-      <CModalHeader>Datasource</CModalHeader>
-      <CModalBody>
-        <Table
-          data={[
-            {
-              key: 'Path',
-              value: props.datasource.rawFilePath,
-            },
-            {
-              key: 'Element ID',
-              value: props.id,
-            },
-          ]}
-        />
-        <CollapseCard>
-          <Table
-            data={[
-              {
-                key: 'Datasource ID',
-                value: props.datasource.id,
-              },
-              {
-                key: 'Status',
-                value: props.state,
-              },
-            ]}
-          />
-        </CollapseCard>
-      </CModalBody>
-    </>
+    <BaseModal isOpen={modalOpened} title="Datasource" onClosed={onClose}>
+      <CoreDataTable
+        tableData={tData}
+        rowHeight="35px"
+        usePagination={false}
+        columns={columns}
+      />
+    </BaseModal>
   )
 }